Relocate got command
