+++ cmd/got/cmd/upsert.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/poka-yoke/spaceflight/mcc/got/got"
+	"github.com/poka-yoke/spaceflight/pkg/got"
 )
 
 var name, typ string
