+++ cmd/got/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/poka-yoke/spaceflight/cmd/got/cmd"
+
+func main() {
+	cmd.Execute()
+}
